Bugfix. Back button.
Change: Added back button to Find match scene (Normal). When player is found, this button is deactivated.
Change: End of turn is now considered if the game is in intern state, as an addition to localId == curId

Fix: At the end of player's turn, now correctly deselects all of the tiles.